requirements.txt
fastapi
uvicorn
gunicorn
jinja2
pytest
requests
httpx
pre-commit
pydantic-settings
python-dotenv
python-multipart
aiofiles
Pillow
pytesseract